rename: VoteSummary 위치 조정 #35
