+++ src/main/java/com/hogu/am_i_hogu/domain/user/dto/response/MyPostItemResponse.java
@@ -1,7 +1,5 @@
 package com.hogu.am_i_hogu.domain.user.dto.response;
 
-import com.hogu.am_i_hogu.domain.user.domain.VoteSummary;
-
 import java.time.LocalDateTime;
 
 public record MyPostItemResponse(
