+++ src/main/java/com/hogu/am_i_hogu/domain/user/dto/response/VoteSummary.java
@@ -1,4 +1,4 @@
-package com.hogu.am_i_hogu.domain.user.domain;
+package com.hogu.am_i_hogu.domain.user.dto.response;
 
 public enum VoteSummary {
     HOGU,
